Reintentos con jitter, y dos count(*) que mentían sobre su tipo
Revisión de la capa de base de datos. Dos arreglos concretos.

1) LA PAUSA DEL REINTENTO ERA FIJA, Y ESO SINCRONIZA LAS EMBESTIDAS.

Un despliegue hornea las nueve páginas en el mismo segundo, así que
caducan casi juntas y se rehacen juntas. Cuando el pooler está lleno
fallan todas a la vez, y con una pausa fija de 400 ms todas volvían a
pedir conexión exactamente 400 ms después: la segunda ola tan simultánea
como la primera. Es lo que se vio a las 09:20 con siete páginas y el
panel.

Ahora la espera lleva azar (backoff con jitter, lo que publicó AWS como
"Exponential Backoff and Jitter" y trata el libro de SRE de Google en el
capítulo de sobrecarga): los que tropiezan juntos vuelven separados.

El azar va DENTRO del presupuesto que ya había —entre la mitad y el total
de la pausa, nunca más— para no tocar el cálculo que ya comprueba
scripts/verificar-plazos.mjs. Una sola verdad sobre los plazos.

2) DOS `count(*)` SIN `::int`.

El propio código lo tiene documentado —un count de Postgres llega como
CADENA, y el tipo decía `number`— y ya causó un fallo real: un aviso del
panel que no salía porque "0" === 0 es falso. Quedaban dos sitios sin la
conversión, en la lista de clientes y en su CSV. Hoy solo se pintaban, así
que no rompían nada; el tipo mentía igual, y el día que alguien ordene por
esa columna, "9" iría después de "10".

Se arregla en el SQL con `::int`, que es como lo hacen las otras cuatro
consultas del proyecto.

Comprobado en local: build entero (12 reglas de plazos, 43+32+14 casos),
paridad 44 filas, 200 navegaciones del panel sin un fallo, y la lista de
clientes y su CSV (200) mirados en el navegador.

+++ src/app/admin/(panel)/clientes/page.tsx
@@ -31,7 +31,7 @@ export default async function PaginaClientes() {
     intentar(
       () => sql<Cliente[]>`
         select p.id, p.full_name, p.phone_e164, m.name as municipality, p.created_at,
-               (select count(*) from app.spins s where s.player_id = p.id) as tiradas
+               (select count(*)::int from app.spins s where s.player_id = p.id) as tiradas
           from app.players p
           join app.municipalities m on m.id = p.municipality_id
          where p.blocked_at is null
